requirements.txt
pandas
numpy
streamlit
Pillow
openai
SpeechRecognition
whisper
Pyaudio
scrubadubdub
